Rename delete container to delete instance
